+++ lib/gds_api/account_api.rb
@@ -127,46 +127,6 @@ class GdsApi::AccountApi < GdsApi::Base
     delete_json("#{endpoint}/api/email-subscriptions/#{CGI.escape(name)}", {}, auth_headers(govuk_account_session))
   end
 
-  # Look up all pages saved by a user in their Account
-  #
-  # @param [String] govuk_account_session Value of the session header
-  #
-  # @return [Hash] containing :saved_pages, an array of single saved page hashes
-  def get_saved_pages(govuk_account_session:)
-    get_json("#{endpoint}/api/saved-pages", auth_headers(govuk_account_session))
-  end
-
-  # Return a single page by unique URL
-  #
-  # @param [String] the path of a page to check
-  # @param [String] govuk_account_session Value of the session header
-  #
-  # @return [Hash] containing :saved_page, a hash of a single saved page value
-  def get_saved_page(page_path:, govuk_account_session:)
-    get_json("#{endpoint}/api/saved-pages/#{CGI.escape(page_path)}", auth_headers(govuk_account_session))
-  end
-
-  # Upsert a single saved page entry in a users account
-  #
-  # @param [String] the path of a page to check
-  # @param [String] govuk_account_session Value of the session header
-  #
-  # @return [Hash] A single saved page value (if sucessful)
-  def save_page(page_path:, govuk_account_session:)
-    put_json("#{endpoint}/api/saved-pages/#{CGI.escape(page_path)}", {}, auth_headers(govuk_account_session))
-  end
-
-  # Delete a single saved page entry from a users account
-  #
-  # @param [String] the path of a page to check
-  # @param [String] govuk_account_session Value of the session header
-  #
-  # @return [GdsApi::Response] A status code of 204 indicates the saved page has been successfully deleted.
-  #                            A status code of 404 indicates there is no saved page with this path.
-  def delete_saved_page(page_path:, govuk_account_session:)
-    delete_json("#{endpoint}/api/saved-pages/#{CGI.escape(page_path)}", {}, auth_headers(govuk_account_session))
-  end
-
 private
 
   def nested_query_string(params)